JUnit Testing Implemented

+++ src/courseManagementSystem/Course.java
@@ -12,14 +12,17 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
 
-public class Course {
+public class Course
+{
 	private String courseID;
 	private String description;
 	private int maxNum;
 	private double cost;
+	private String[] courseIDs = { "001", "002", "003", "004", "005" };
 
 	// constructor
-	public Course(String courseID, String courseName, int maxStudents, double coursePrice) {
+	public Course(String courseID, String courseName, int maxStudents, double coursePrice)
+	{
 
 		this.courseID = courseID;
 		this.description = courseName;
@@ -28,15 +31,18 @@ public class Course {
 	}
 
 	// Writer to write details to text file
-	public void writeUsingFileWriter(String courseID, String description, int maxNum, double cost)
-			throws FileNotFoundException {
+	public boolean writeUsingFileWriter(String courseID, String description, int maxNum, double cost)
+			throws FileNotFoundException
+	{
 		// find users current directory
 		File file = new File(System.getProperty("user.dir"));
 		// then access the course details text file
 		String path = file.getAbsolutePath() + "\\src\\courseDetails.txt";
 		File created = new File(path);
-		try {
-			if (created.exists() == false) {
+		try
+		{
+			if (created.exists() == false)
+			{
 				System.out.println("We had to make a new file.");
 				created.createNewFile();
 			}
@@ -44,56 +50,89 @@ public class Course {
 			System.out.println("Record Successfully Updated!!");
 			out.append("{" + courseID + "," + description + "," + maxNum + "," + cost + ",}\n");
 			out.close();
-		} catch (IOException e) {
+		} catch (IOException e)
+		{
 			System.out.println("COULD NOT UPDATE RECORD!!");
+			return false;
 		}
+		return true;
 
 	}
 
 	// delete line in text file
-	public void deleteLine() {
-		try {
+	public boolean deleteLine()
+	{
+		boolean exist = false;
+		try
+		{
 			BufferedReader file = new BufferedReader(new FileReader("src/courseDetails.txt"));
 			String line;
 			String input = "";
 			@SuppressWarnings("resource")
 			Scanner scan = new Scanner(System.in);
 			System.out.println("Enter course ID to change:");
-			courseID = scan.nextLine();
+			String checkCourseID = scan.nextLine();
 			System.out.println("Enter new course cost:");
 			cost = scan.nextDouble();
-			while ((line = file.readLine()) != null) {
+			while ((line = file.readLine()) != null)
+			{
 				// check for line with course ID
-				if (line.contains(courseID)) {
+				if (line.contains(checkCourseID))
+				{
 					// read text file and assign line to variables for rewriting
 					String[] checkCredentials = line.split("}");
 					String[] jobSearch = checkCredentials[0].split(",");
 					courseID = jobSearch[0].replace("{", "");
-					description = jobSearch[1];
-					maxNum = Integer.parseInt(jobSearch[2]);
-					line = "";
+					if (checkCourseID.equals(courseID))
+					{
+						description = jobSearch[1];
+						maxNum = Integer.parseInt(jobSearch[2]);
+						line = "";
+						exist = true;
+
+					}
+
 				}
+
 				input += line + '\n';
+
+			}
+			if (!checkCourseID.equals(courseID))
+			{
+				System.out.println("Course does not exist!!!");
+				exist = false;
+				file.close();
+
 			}
 
-			// assign where to write new variables
-			FileOutputStream File = new FileOutputStream("src/courseDetails.txt");
-			File.write(input.getBytes());
-			writeUsingFileWriter(courseID, description, maxNum, cost);
-			file.close();
-			File.close();
-		} catch (Exception e) {
+			if (exist == true)
+			{
+				// assign where to write new variables
+				FileOutputStream File = new FileOutputStream("src/courseDetails.txt");
+				File.write(input.getBytes());
+				writeUsingFileWriter(courseID, description, maxNum, cost);
+				file.close();
+				File.close();
+				return true;
+			}
+
+		} catch (Exception e)
+		{
 			System.out.println("Problem reading file.");
+			return false;
 		}
+		return false;
 	}
 
-	public void viewCourseDetails() {
+	public boolean viewCourseDetails()
+	{
 		@SuppressWarnings("resource")
 		Scanner scan = new Scanner(System.in);
 		File file = new File(System.getProperty("user.dir"));
 		String path = file.getAbsolutePath() + "\\src\\courseDetails.txt";
 		String fileName = path;
-		try {
+		try
+		{
 			// file reader to read the fileName variable above
 			FileReader filereader = new FileReader(fileName);
 
@@ -107,8 +146,10 @@ public class Course {
 			System.out.printf("%10s %30s %20s %10s", "Course ID", "Course Name", "Max Slots", "Price $");
 			System.out.println();
 			System.out.println("-------------------------------------------------------------------------------------");
-			while ((lineInput = bufferedReader.readLine()) != null) {
-				if (lineInput.length() > 0) {
+			while ((lineInput = bufferedReader.readLine()) != null)
+			{
+				if (lineInput.length() > 0)
+				{
 					list.add(lineInput);
 					// array to split text on line to get user name and password
 					String[] checkCredentials = lineInput.split(" }");
@@ -128,42 +169,71 @@ public class Course {
 			// close buffered reader
 			System.out.print("1. Edit course fee\n2. View Specific Course Numbers\n3. Back to Menu\n");
 			int input = scan.nextInt();
-			if (input == 1) {
+			if (input == 1)
+			{
 				deleteLine();
-			} else if (input == 2) {
+			}
+			else if (input == 2)
+			{
 				printCourseDetails();
-			} else if (input <= 0 || input > 3) {
+			}
+			else if (input <= 0 || input > 3)
+			{
 				System.out.println("Invalid input");
 				viewCourseDetails();
 			}
 			bufferedReader.close();
-		} catch (FileNotFoundException ex) {
+		} catch (FileNotFoundException ex)
+		{
 			System.out.println("Unable to open file '" + fileName + "'");
+			return false;
 		}
 		// catch exception if IOException
-		catch (IOException ex) {
+		catch (IOException ex)
+		{
 			System.out.println("Error reading file '" + fileName + "'");
+			return false;
 
 		}
+		return true;
 
 	}
 
-	public void printCourseDetails() {
+	public boolean printCourseDetails()
+	{
 
 		@SuppressWarnings("resource")
 		Scanner scan = new Scanner(System.in);
 		File file = new File(System.getProperty("user.dir"));
 		String path = file.getAbsolutePath() + "\\src\\student_course_details.txt";
 		String fileName = path;
-		try {
+		try
+		{
 			// file reader to read the fileName variable above
 			FileReader filereader = new FileReader(fileName);
 
 			BufferedReader bufferedReader = new BufferedReader(filereader);
 			// variable for lines in the file
 			String lineInput;
-			System.out.println("Enter Course ID to see details");
-			String courseIdDetails = scan.nextLine();
+			String courseIdDetails = "";
+			boolean exist = false;
+			do
+			{
+				System.out.println("Enter Course ID to see details");
+				courseIdDetails = scan.nextLine();
+				for (int i = 0; i < courseIDs.length; i++)
+				{
+					if (courseIdDetails.equals(courseIDs[i]))
+					{
+						exist = true;
+						break;
+					}
+
+				}
+				if (exist == false)
+					System.out.println("Course ID does not exist, please try again\n");
+			} while (exist == false);
+
 			String studentID = null;
 			double cost = 0;
 			double totalCost = 0;
@@ -173,8 +243,10 @@ public class Course {
 			System.out.printf("%10s %30s", "Student ID", "Fee Paid $");
 			System.out.println();
 			System.out.println("-------------------------------------------------------------------------------------");
-			while ((lineInput = bufferedReader.readLine()) != null) {
-				if (lineInput.length() > 0) {
+			while ((lineInput = bufferedReader.readLine()) != null)
+			{
+				if (lineInput.length() > 0)
+				{
 					list.add(lineInput);
 					// array to split text on line to get user name and password
 					String[] checkCredentials = lineInput.split(" }");
@@ -184,13 +256,16 @@ public class Course {
 					cost = Double.parseDouble(individualRecord[2]);
 					// if correct course, add total amount, and
 					// print out student ID and cost
-					if (courseID.equals(courseIdDetails)) {
+					if (courseID.equals(courseIdDetails))
+					{
+
 						totalCost += cost;
 						System.out.format("%10s %30.2f  ", studentID, cost);
 						System.out.println();
 					}
 				}
 			}
+
 			System.out.println();
 			System.out.format("%15s %25.2f  ", "Total Amount $:", totalCost);
 			System.out.println();
@@ -198,22 +273,27 @@ public class Course {
 			System.out.println();
 			bufferedReader.close();
 
-		} catch (FileNotFoundException ex) {
+			
+		} catch (FileNotFoundException ex)
+		{
 			System.out.println("Unable to open file '" + fileName + "'");
 		}
 		// catch exception if IOException
-		catch (IOException ex) {
+		catch (IOException ex)
+		{
 			System.out.println("Error reading file '" + fileName + "'");
 		}
+		return true;
 	}
 
-	public void printStudentCourseDetails(String checkID) {
+	public boolean printStudentCourseDetails(String checkID)
+	{
 
-		@SuppressWarnings("resource")
 		File file = new File(System.getProperty("user.dir"));
 		String path = file.getAbsolutePath() + "\\src\\student_course_details.txt";
 		String fileName = path;
-		try {
+		try
+		{
 			// file reader to read the fileName variable above
 			FileReader filereader = new FileReader(fileName);
 
@@ -224,15 +304,18 @@ public class Course {
 			int totalCourses = 0;
 			String courseName = "";
 			List<String> list = new ArrayList<String>();
-			System.out.println("\nCourses Taken:");
+			
+			System.out.println("\nCourses Taken by " + checkID);
 			// read through document while there is a new line in the file
 			System.out
 					.println("-------------------------------------------------------------------------------------\n");
 			System.out.printf("%10s %30s", "Course ID", "Course Name\n");
 			// System.out.println();
 			System.out.println("-------------------------------------------------------------------------------------");
-			while ((lineInput = bufferedReader.readLine()) != null) {
-				if (lineInput.length() > 0) {
+			while ((lineInput = bufferedReader.readLine()) != null)
+			{
+				if (lineInput.length() > 0)
+				{
 					list.add(lineInput);
 					// array to split text on line to get user name and password
 					String[] checkCredentials = lineInput.split(" }");
@@ -253,57 +336,74 @@ public class Course {
 
 					// if correct course, add total amount, and
 					// print out student ID and cost
-					if (studentID.equals(checkID)) {
+					if (studentID.equals(checkID))
+					{
 						totalCourses++;
 						System.out.format("%10s %30s  ", courseID, courseName);
 						System.out.println();
+
 					}
+
 				}
+
 			}
-			System.out.println();
-			System.out.format("%15s %25d  ", "Total Courses:", totalCourses);
-			System.out.println();
-			System.out.println("-------------------------------------------------------------------------------------");
-			System.out.println();
+
 			bufferedReader.close();
 
-		} catch (FileNotFoundException ex) {
+			if (!studentID.equals(checkID))
+				return false;
+
+			else
+			{
+				System.out.println();
+				System.out.format("%15s %25d  ", "Total Courses:", totalCourses);
+				System.out.println();
+				System.out.println(
+						"-------------------------------------------------------------------------------------");
+				System.out.println();
+			}
+		} catch (FileNotFoundException ex)
+		{
 			System.out.println("Unable to open file '" + fileName + "'");
+			return false;
 		}
 		// catch exception if IOException
-		catch (IOException ex) {
+		catch (IOException ex)
+		{
 			System.out.println("Error reading file '" + fileName + "'");
+			return false;
 		}
+		return true;
 	}
 
-	public void calculate()
+	public boolean calculate()
 	{
-		String[] courseIDs = { "001", "002", "003", "004", "005" };
+
 		for (int i = 0; i < courseIDs.length; i++)
 		{
-		@SuppressWarnings("resource")
-		File file = new File(System.getProperty("user.dir"));
-		String path = file.getAbsolutePath() + "\\src\\student_course_details.txt";
-		String fileName = path;
-		try
-		{
-			// file reader to read the fileName variable above
-			FileReader filereader = new FileReader(fileName);
 
-			BufferedReader bufferedReader = new BufferedReader(filereader);
-			// variable for lines in the file
-			String lineInput;
-			String courseID = null;
-			String courseName = "";
-		
-			List<String> list = new ArrayList<String>();
-		
+			File file = new File(System.getProperty("user.dir"));
+			String path = file.getAbsolutePath() + "\\src\\student_course_details.txt";
+			String fileName = path;
+			try
+			{
+				// file reader to read the fileName variable above
+				FileReader filereader = new FileReader(fileName);
+
+				BufferedReader bufferedReader = new BufferedReader(filereader);
+				// variable for lines in the file
+				String lineInput;
+				String courseID = null;
+				String courseName = "";
+
+				List<String> list = new ArrayList<String>();
+
 				double cost;
 				double runningCost = 0;
 				double profit = 0;
 				double totalCost = 0;
 				int numOfStudents = 0;
-				
+
 				while ((lineInput = bufferedReader.readLine()) != null)
 				{
 					if (lineInput.length() > 0)
@@ -315,30 +415,35 @@ public class Course {
 						String[] checkCredentials = lineInput.split(" }");
 						String[] individualRecord = checkCredentials[0].split(",");
 						courseID = individualRecord[0].replace("{", "");
-						//System.out.println("Course ID:" +courseID);
-						
+						// System.out.println("Course ID:" +courseID);
+
 						if (courseIDs[i].equals(courseID))
 						{
 							cost = Double.parseDouble(individualRecord[2]);
 							totalCost += cost;
 							numOfStudents++;
-							if (courseIDs[i] == "001") {
+							if (courseIDs[i] == "001")
+							{
 								courseName = "Italian Cooking";
 								runningCost = 1000;
 							}
-							if (courseIDs[i] == "002") {
+							if (courseIDs[i] == "002")
+							{
 								courseName = "Seafood Cooking";
 								runningCost = 1000;
-							}							
-							if (courseIDs[i] == "003") {
+							}
+							if (courseIDs[i] == "003")
+							{
 								courseName = "Sewing";
 								runningCost = 100 * numOfStudents;
 							}
-							if (courseIDs[i] == "004" ) {
+							if (courseIDs[i] == "004")
+							{
 								courseName = "Creative Writing";
 								runningCost = 800;
 							}
-							if (courseIDs[i] == "005") {
+							if (courseIDs[i] == "005")
+							{
 								courseName = "Business Writing";
 								runningCost = 600;
 							}
@@ -349,23 +454,28 @@ public class Course {
 				}
 				bufferedReader.close();
 				System.out.println();
-				System.out.format("%10s %5s %17s %15s %5s %10s %10.2f %10s %10.2f %10s %10.2f", "Course:", courseIDs[i], courseName, "Students:", numOfStudents,"Cost:", runningCost, "Income:", totalCost,"Profit:",profit);
+				System.out.format("%10s %5s %17s %15s %5s %10s %10.2f %10s %10.2f %10s %10.2f", "Course:", courseIDs[i],
+						courseName, "Students:", numOfStudents, "Cost:", runningCost, "Income:", totalCost, "Profit:",
+						profit);
 				System.out.println();
 				System.out.println(
 						"----------------------------------------------------------------------------------------------------------------------------");
 				System.out.println();
 			}
-		
-		catch (FileNotFoundException ex)
-		{
-			System.out.println("Unable to open file '" + fileName + "'");
+
+			catch (FileNotFoundException ex)
+			{
+				System.out.println("Unable to open file '" + fileName + "'");
+				return false;
+			}
+			// catch exception if IOException
+			catch (IOException ex)
+			{
+				System.out.println("Error reading file '" + fileName + "'");
+				return false;
+			}
 		}
-		// catch exception if IOException
-		catch (IOException ex)
-		{
-			System.out.println("Error reading file '" + fileName + "'");
-		}
-	}
+		return true;
 	}
 
 }
